Format numbers in right side of transaction windows
