+++ ui/app/components/token-balance/token-balance.component.js
@@ -2,6 +2,8 @@ import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
 import classnames from 'classnames'
 
+import { formatTokenAmount } from '../../helpers/formatter-numbers.util'
+
 export default class TokenBalance extends PureComponent {
   static propTypes = {
     token: PropTypes.object,
@@ -13,9 +15,10 @@ export default class TokenBalance extends PureComponent {
     const { className, withSymbol, token } = this.props
     const { symbol, string } = token
 
+
     return (
       <div className={classnames('hide-text-overflow', className)}>
-        {string + (withSymbol ? ` ${symbol}` : '')}
+        {formatTokenAmount(string || '0') + (withSymbol ? ` ${symbol}` : '')}
       </div>
     )
   }
