+++ ui/app/helpers/formatter-numbers.util.js
@@ -1,0 +1,14 @@
+// Amount comes in as a string
+const formatTokenAmount = (amount) => {
+  try {
+    const value = parseFloat(amount)
+    const formatted = value.toLocaleString()
+    return formatted
+  } catch (e) {
+    return amount
+  }
+}
+
+export {
+  formatTokenAmount,
+}
